refactor: migrate Users tests
